Change INTEGER to BIGINT for tables
Tweet IDs are huge.

+++ models/log.js
@@ -1,12 +1,12 @@
-import {INTEGER} from 'sequelize';
+import {BIGINT} from 'sequelize';
 import sequelize from './';
 
 const Log = sequelize.define('log', {
-  id: {type: INTEGER, autoIncrement: true, primaryKey: true},
-  from: {type: INTEGER, allowNull: true, defaultValue: null},
-  to: {type: INTEGER, allowNull: true, defaultValue: null},
-  resultFrom: {type: INTEGER, allowNull: true, defaultValue: null},
-  resultTo: {type: INTEGER, allowNull: true, defaultValue: null}
+  id: {type: BIGINT, autoIncrement: true, primaryKey: true},
+  from: {type: BIGINT, allowNull: true, defaultValue: null},
+  to: {type: BIGINT, allowNull: true, defaultValue: null},
+  resultFrom: {type: BIGINT, allowNull: true, defaultValue: null},
+  resultTo: {type: BIGINT, allowNull: true, defaultValue: null}
 }, {
   classMethods: {
     latest() {
